restructure identity popover
