+++ utils/format/time.ts
@@ -1,5 +1,10 @@
+import {
+  addSeconds,
+  formatDistanceToNow,
+  formatDuration,
+  intervalToDuration,
+} from 'date-fns'
 import { parseDate } from '../datetime'
-import { intervalToDuration, formatDuration, addSeconds } from 'date-fns'
 
 export const formatSecondsToDuration = (seconds: number) => {
   const duration = intervalToDuration({
@@ -28,4 +33,8 @@ export const formatSecondsToDuration = (seconds: number) => {
 
 export const endDate = (seconds: number): string => {
   return parseDate(addSeconds(new Date(), seconds))
+}
+
+export const formatToNow = (date: Date): string => {
+  return formatDistanceToNow(new Date(date), { addSuffix: true })
 }
